@@ -4,19 +4,24 @@
   "version": "1.0.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "bootstrap": "^4.0.0",
-    "gatsby": "^1.9.207",
-    "gatsby-link": "^1.6.37",
-    "gatsby-plugin-react-helmet": "^2.0.5",
-    "gatsby-source-filesystem": "^1.5.25",
-    "gatsby-transformer-json": "^1.0.16",
-    "jquery": "^3.3.1",
+    "aphrodite": "^2.4.0",
+    "bootstrap": "^4.5.2",
+    "gatsby": "^2.24.63",
+    "gatsby-link": "^2.4.14",
+    "gatsby-plugin-react-helmet": "^3.3.11",
+    "gatsby-source-filesystem": "^2.3.30",
+    "gatsby-transformer-json": "^2.4.12",
+    "jquery": "^3.5.1",
     "popper.js": "^1.12.9",
-    "react-helmet": "^5.2.0",
-    "react-scrollspy": "^3.3.5",
-    "react-slick": "^0.23.1",
-    "react-transition-group": "^1.2.1",
-    "reactstrap": "^5.0.0-beta",
+    "react": "^16.13.1",
+    "react-animate-on-scroll": "^2.1.5",
+    "react-animations": "^1.0.0",
+    "react-dom": "^16.13.1",
+    "react-helmet": "^6.1.0",
+    "react-scrollspy": "^3.4.3",
+    "react-slick": "^0.27.11",
+    "react-transition-group": "^4.4.1",
+    "reactstrap": "^8.6.0",
     "slick-carousel": "^1.8.1"
   },
   "keywords": [
@@ -32,8 +37,8 @@
     "eslint": "eslint . --ext \".js,.jsx\"; exit 0"
   },
   "devDependencies": {
-    "eslint": "^4.18.2",
-    "eslint-plugin-react": "^7.7.0",
-    "prettier": "^1.10.2"
+    "eslint": "^7.9.0",
+    "eslint-plugin-react": "^7.20.6",
+    "prettier": "^2.1.2"
   }
 }
