@@ -1,10 +1,11 @@
 module.exports = {
   siteMetadata: {
     title: 'CIDEMIPYME',
-    slogan: 'Hacemos Negocio su Negocio',
+    slogan: 'Hacemos Negocio Tu Negocio',
     fbcide: 'https://www.facebook.com/cidemipyme',
     fbcce: 'https://www.facebook.com/circulodecrecimiento',
     fbcidetalento: 'https://www.facebook.com/cidetalento',
+    linkedin: 'https://www.linkedin.com/company/cidemipyme/'
   },
   plugins: [
     {
