@@ -6,8 +6,8 @@
 
 const path = require('path')
 
-exports.modifyWebpackConfig = ({ config, env }) => {
-  config.merge({
+exports.onCreateWebpackConfig = ({ stage, actions }) => {
+  actions.setWebpackConfig({
     resolve: {
       alias: {
         images: path.resolve(__dirname, 'src/assets/images'),
@@ -16,8 +16,9 @@ exports.modifyWebpackConfig = ({ config, env }) => {
         fonts: path.resolve(__dirname, 'src/assets/fonts'),
         css: path.resolve(__dirname, 'src/assets/css'),
       }
+    },
+    externals: {
+      jquery: 'jQuery',
     }
-  });
-
-  return config;
+  })
 }
