@@ -5,6 +5,7 @@ import methodology_cidetalento from 'images/metodologia_cidetalento.png';
 import methodology_growth_strategy from 'images/metodologia_estrategia_crecimiento.png';
 import PropTypes from 'prop-types';
 import React from 'react';
+import { graphql } from 'gatsby';
 import {
   Col,
   Container,
@@ -67,7 +68,7 @@ class MethodologiesSection extends React.Component {
     }
 
     return (
-      <section id="section-methodologies" className="pb_section bg-light">
+      <section id="section-methodologies" className="pb_section">
         <Container>
           <Row>
             <Col md={12}>
