@@ -1,91 +1,151 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
+import React, { useState, useRef } from 'react';
+import { graphql } from 'gatsby';
 import {
   Container,
   Row,
   Col,
 } from 'reactstrap';
-import cideServices from 'images/cide_services.png';
+import Slider from "react-slick";
+import SlideCard from 'components/SlideCard.jsx';
+import Scrollspy from 'react-scrollspy';
+
 
 const workKeyPrefix = 'workSection_';
+const serviceSlideKeyPrefix = 'serviceSlide_';
 
-const WorkInnerItemPropType = {
-  icon: PropTypes.string.isRequired,
-  items: PropTypes.arrayOf(PropTypes.string).isRequired,
-}
-
-const WorkPropType = {
-  title: PropTypes.string.isRequired,
-  work_items: PropTypes.arrayOf(PropTypes.shape(WorkInnerItemPropType)).isRequired
-}
-
-class WorkSection extends React.Component {
-  createWorkCard(icon, items, key) {
-    const textItems = items.map((text, index) => (
-      <li key={`${key}_text_${index}`}>{text}</li>
-    ));
-    return (
-      <Col md={6} key={`${key}_work_card`}>
-        <div className="media pb_media_v1 mb-5">
-          <div className="icon border border-gray rounded-circle d-flex mr-3 display-4 text-primary">
-            <i className={`flaticon ${icon}`}/>
-          </div>
-          <div className="media-body">
-            <ul className="pb_font-14">
-              {textItems}
-            </ul>
-          </div>
+const ServiceWithIconItem = (props) => {
+  const {
+    title,
+    icon,
+    content,
+    offset,
+    onSelect
+  } = props;
+  return (
+    <Col lg={{ size: 4, offset }} className="service-work-section-unit">
+      <Row>
+        <div className="d-flex mr-3 display-4 icon-primary-color strategy-icon">
+          <i className={`flaticon ${icon}`} />
         </div>
-      </Col>
-    );
-  }
+      </Row>
+      <Row>
+        <div className="media-body text-center">
+          <h3 className="mt-0 pb_font-22">
+            {title}
+          </h3>
+          <p className="pb_font-15 mb-0">
+            {content}
+          </p>
+          <Scrollspy 
+            items={['slider-services']}
+            className="pl-0 ml-0 pb_font-15"
+            offset={100}>
+            <a href="#slider-services" onClick={onSelect}>Ver Más</a>
+          </Scrollspy>
+        </div>
+      </Row>
+    </Col>
+  );
+}
 
-  render() {
-    const cards = this.props.data.work_items.map((item, index) => (
-      this.createWorkCard(
-        item.icon,
-        item.items,
-        `${workKeyPrefix}card_${index}`
-      )
+const SlickArrow = ({ currentSlide, slideCount, side, ...props }) => (
+  <span {...props}><i className={`flaticon ion-ios-arrow-${side}`} /></span>
+);
+
+const WorkSection = (props) => {
+  const { title, work_items: services } = props;
+  const sliderRef = useRef(null);
+  const [sliderIndex, setSliderIndex] = useState(0);
+
+  const cardsFirstRow = services.slice(0, 3).map((item, index) =>
+    <ServiceWithIconItem
+      {...item}
+      key={`${workKeyPrefix}1card_${index}`}
+      offset={0}
+      onSelect={() => sliderRef.current.slickGoTo(index)} />
+  );
+
+  const cardsSecondRow = services.slice(3).map((item, index) =>
+    <ServiceWithIconItem
+      {...item}
+      key={`${workKeyPrefix}2card_${index}`}
+      offset={1}
+      onSelect={() => sliderRef.current.slickGoTo(index + 3)} />
+  );
+
+  const sliderSettings = {
+    dots: true,
+    infinite: true,
+    slidesToShow: 1,
+    slidesToScroll: 1,
+    arrows: true,
+    nextArrow: <SlickArrow side="right" />,
+    prevArrow: <SlickArrow side="left" />,
+    beforeChange: (_current, next) => setSliderIndex(next)
+  };
+  const slides = services.map(
+    (service, index) => (
+      <SlideCard
+        cssClassImage={service.slide.css_class}
+        icon={service.icon}
+        title={service.title}
+        content={service.slide.content}
+        key={`${serviceSlideKeyPrefix}${index}_`}
+        keyPrefix={`${serviceSlideKeyPrefix}${index}_`} />
     ));
 
-    return (
-      <section id="section-work" className="pb_section">
-        <Container>
-          <Row>
-            <Col>
-              <h2 className="mt-0 heading-border-top mb-3 font-weight-normal section-header">
-                {this.props.data.title}
-              </h2>
-            </Col>
-          </Row>
-          <Row>
-            {cards}
-          </Row>
-          <Row>
-            <Col sm={12} md={{size: 8, offset:2}}>
-              <img className="services-image" alt="CIDEMiPyMe Diagrama de Servicios" src={cideServices}/>
-            </Col>
-          </Row>
-        </Container>
-      </section>
-    );
-  }
+  return (
+    <>
+    <section id="section-work" className="pb_section">
+      <Container>
+        <Row>
+          <Col sm={10}>
+            <h2 className="mt-0 heading-border-top mb-3 font-weight-normal section-header">
+              {title}
+            </h2>
+          </Col>
+        </Row>
+        <Row>
+          {cardsFirstRow}
+        </Row>
+        <Row>
+          {cardsSecondRow}
+        </Row>
+      </Container>
+    </section>
+    <section className="pb_section bg-light">
+      <Container>
+      <Row>
+      <div id="slider-services" />
+      <Col lg={12} md={12} sm={0}>
+        <Slider
+          ref={sliderRef} {...sliderSettings}
+          className="single-item pb_slide_v2">
+          {slides}
+        </Slider>
+      </Col>
+    </Row>
+      </Container>
+    
+    </section>
+    </>
+  );
 }
 
 export default WorkSection;
 
-WorkSection.propTypes = {
-  data: PropTypes.shape(WorkPropType).isRequired
-}
-
-export const GraphQlWorkSectionFragment = graphql `
+export const GraphQlWorkSectionFragment = graphql`
   fragment WorkSectionFragment on InformationJson {
     title
     work_items {
       icon
-      items
+      title
+      content
+      image
+      slide {
+        css_class
+        content
+      }
     }
   }
 `;
